Fixing tests in bigint and util
